+++ containers/daap-authorizer/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0]
+
+### Changed
+
+- Use custom logging library image
+- Use custom logging library
+
 ## [1.0.1] 2023-08-30
 
 ### Changed